+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fixed issue in `Named` where old code wasn't deleted, causing some named properties to be completely omitted.
 * Improved efficiency of `Named` by making it so `get` no longer copied the entire dictionary repeatedly while looking for the property.
 * Fixed typo in `Attachment` that caused embedded msg files to still regenerate the `Named` instance even though they should have been using the parent's.
+* Updated fields in `MSGFile` to use enums.
+* Added additional properties to `MSGFile`.
 
 **v0.32.0**
 * [[TeamMsgExtractor #217](https://github.com/TeamMsgExtractor/msg-extractor/issues/217)] Redesigned named properties to be much more efficient. All in all, load times for MSG files are significantly improved all around.